Split up CaseExpression
